requirements.txt
azure-functions
pydub
requests
